+++ packages/r3bl-auto-insert-copyright/src/configuration.ts
@@ -24,6 +24,7 @@
 import * as vscode from 'vscode';
 import { Copyright } from './copyright/copyright';
 import { Apache2 } from './copyright/licenses/apache2';
+import { Apache2OneLine } from './copyright/licenses/apache2-one-line';
 import { Gpl } from './copyright/licenses/gpl3';
 import { Mit } from './copyright/licenses/mit';
 import { Proprietary } from './copyright/licenses/proprietary';
@@ -59,6 +60,8 @@ export function getCopyright(): Copyright {
 
   if (selectedLicense === 'Apache2') {
     return new Apache2();
+  } else if (selectedLicense === 'Apache2OneLine') {
+    return new Apache2OneLine();
   } else if (selectedLicense === 'MIT') {
     return new Mit();
   } else if (selectedLicense === 'GPL3') {