+++ packages/r3bl-auto-insert-copyright/src/copyright/copyrightService.ts
@@ -1,18 +1,24 @@
 /*
- *   Copyright (c) 2019 Ford Motor Company
+ *   Copyright (c) 2024 R3BL LLC
  *   All rights reserved.
-
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
-
- *   http://www.apache.org/licenses/LICENSE-2.0
-
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
+ *
+ *   Permission is hereby granted, free of charge, to any person obtaining a copy
+ *   of this software and associated documentation files (the "Software"), to deal
+ *   in the Software without restriction, including without limitation the rights
+ *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *   copies of the Software, and to permit persons to whom the Software is
+ *   furnished to do so, subject to the following conditions:
+ *
+ *   The above copyright notice and this permission notice shall be included in all
+ *   copies or substantial portions of the Software.
+ *
+ *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *   SOFTWARE.
  */
 
 import * as vscode from 'vscode';
@@ -41,16 +47,38 @@ export function handleManualCopyrightCheck(editor: vscode.TextEditor | undefined
   }
 }
 
+/**
+ * Checks if a document already contains a copyright header.
+ * 
+ * Supports detection of both single-line and multi-line copyright formats:
+ * - Single-line: Apache2OneLine (// Copyright...)
+ * - Multi-line: MIT, Apache2, GPL3 (/* ... * Copyright...)
+ * 
+ * @param document The VSCode text document to check
+ * @returns true if a copyright header is detected, false otherwise
+ */
 function hasCopyright(document: vscode.TextDocument): Boolean {
   if (isNewDocument(document)) {
     return false;
   }
 
-  const copyrightLine = document.lineAt(1);
-  return (
-    !copyrightLine.isEmptyOrWhitespace &&
-    copyrightLine.text.includes('Copyright')
-  );
+  // Check first line (single-line copyright like Apache2OneLine: "// Copyright...")
+  const firstLine = document.lineAt(0);
+  if (!firstLine.isEmptyOrWhitespace && firstLine.text.includes('Copyright')) {
+    return true;
+  }
+
+  // Check second line (multi-line copyright like MIT/Apache2/GPL3: "/* ... * Copyright...")
+  // Note: Multi-line licenses have "Copyright" on line 3-4, but we check line 1 which
+  // contains the block comment structure indicating a copyright header is present
+  if (document.lineCount > 1) {
+    const secondLine = document.lineAt(1);
+    if (!secondLine.isEmptyOrWhitespace && secondLine.text.includes('Copyright')) {
+      return true;
+    }
+  }
+
+  return false;
 }
 
 function insertCopyright(editor: vscode.TextEditor) {