+++ packages/r3bl-auto-insert-copyright/src/copyright/licenses/apache2.ts
@@ -21,8 +21,6 @@
  *   SOFTWARE.
  */
 
-
-
 'use strict';
 
 import { Copyright } from '../copyright';