+++ packages/r3bl-auto-insert-copyright/src/copyright/licenses/mit.ts
@@ -1,5 +1,5 @@
 /*
- *   Copyright (c) 2022 
+ *   Copyright (c) 2022
  *   All rights reserved.
  */
 'use strict'
@@ -13,8 +13,9 @@ export class Mit extends Copyright {
 
   public header(): string {
     let template = `/*
+ *   MIT License
+ *
  *   Copyright (c) ${this.year} ${this.author}
- *   All rights reserved.
  *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
  *   of this software and associated documentation files (the "Software"), to deal
@@ -22,10 +23,10 @@ export class Mit extends Copyright {
  *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  *   copies of the Software, and to permit persons to whom the Software is
  *   furnished to do so, subject to the following conditions:
- * 
+ *
  *   The above copyright notice and this permission notice shall be included in all
  *   copies or substantial portions of the Software.
- * 
+ *
  *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE